Renames existing callbacks into consistent naming convention
All callbacks are now named foo_callback.h. So, that they are more easily
recognized as being related.
