+++ geometry/proximity/distance_to_point_callback.h
