+++ geometry/proximity/find_collision_candidates_callback.h
